use int for index

+++ ecs/query_gen.go
@@ -49,11 +49,11 @@ func (q *Query0) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query0) EntityAt(index uint32) Entity {
+func (q *Query0) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAt(&q.world.storage, q.filter, q.relations, index)
+		return entityAt(&q.world.storage, q.filter, q.relations, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -183,11 +183,11 @@ func (q *Query1[A]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query1[A]) EntityAt(index uint32) Entity {
+func (q *Query1[A]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -319,11 +319,11 @@ func (q *Query2[A, B]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query2[A, B]) EntityAt(index uint32) Entity {
+func (q *Query2[A, B]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -460,11 +460,11 @@ func (q *Query3[A, B, C]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query3[A, B, C]) EntityAt(index uint32) Entity {
+func (q *Query3[A, B, C]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -604,11 +604,11 @@ func (q *Query4[A, B, C, D]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query4[A, B, C, D]) EntityAt(index uint32) Entity {
+func (q *Query4[A, B, C, D]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -751,11 +751,11 @@ func (q *Query5[A, B, C, D, E]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query5[A, B, C, D, E]) EntityAt(index uint32) Entity {
+func (q *Query5[A, B, C, D, E]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -901,11 +901,11 @@ func (q *Query6[A, B, C, D, E, F]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query6[A, B, C, D, E, F]) EntityAt(index uint32) Entity {
+func (q *Query6[A, B, C, D, E, F]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -1054,11 +1054,11 @@ func (q *Query7[A, B, C, D, E, F, G]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query7[A, B, C, D, E, F, G]) EntityAt(index uint32) Entity {
+func (q *Query7[A, B, C, D, E, F, G]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
@@ -1210,11 +1210,11 @@ func (q *Query8[A, B, C, D, E, F, G, H]) Count() int {
 // Do not use this to iterate a query! Use [Query.Next] instead.
 //
 // Panics if the index is out of range, as indicated by [Query.Count].
-func (q *Query8[A, B, C, D, E, F, G, H]) EntityAt(index uint32) Entity {
+func (q *Query8[A, B, C, D, E, F, G, H]) EntityAt(index int) Entity {
 	if q.cache == nil {
-		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, index)
+		return entityAtComponent(&q.world.storage, q.filter, q.relations, q.rareComp, uint32(index))
 	}
-	return entityAtCache(&q.world.storage, q.cache, q.relations, index)
+	return entityAtCache(&q.world.storage, q.cache, q.relations, uint32(index))
 }
 
 // Close closes the Query and unlocks the world.
